Fix Mixins For Rehydration
There was an issue with the mixin pattern that you are supposed to use in TS 2.2, however this introduced type errors that got past CI. As a work around we can just use a sub-class for the rehydration tests.
