+++ packages/@glimmer/test-helpers/lib/abstract-test-case.ts
@@ -15,8 +15,6 @@ import {
   regex,
   BasicComponent
 } from "../lib/environment";
-import { NodeDOMTreeConstruction } from "@glimmer/node";
-import * as SimpleDOM from "simple-dom";
 
 export const OPEN: { marker: "open-block" } = { marker: "open-block" };
 export const CLOSE: { marker: "close-block" } = { marker: "close-block" };
@@ -486,93 +484,6 @@ export class RenderTests extends AbstractRenderTest {
   }
 }
 
-export class RehydrationTests extends RenderTests {
-  serialized: string;
-  setupServer(template: string = this.rawTemplate) {
-    let doc = new SimpleDOM.Document();
-    let env = new TestEnvironment({
-      document: doc,
-      appendOperations: new NodeDOMTreeConstruction(doc)
-    });
-    this.setup({ template, env });
-  }
-
-  setupClient(template: string = this.rawTemplate) {
-    let env = new TestEnvironment();
-    let div = document.createElement("div");
-
-    expect(this.serialized, "Should have serialized HTML from `this.renderServerSide()`");
-
-    div.innerHTML = this.serialized;
-    this.element = div;
-    this.setup({ template, env });
-  }
-
-  setup({ template, context, env }: { template: string; context?: Dict<Opaque>; env?: TestEnvironment }) {
-    if (env) this.env = env;
-    this.template = this.compile(template);
-    if (context) this.setProperties(context);
-  }
-
-  assertServerOutput(..._expected: Content[]) {
-    let serialized = this.serialize();
-    equalTokens(serialized, content([OPEN, ..._expected, CLOSE]));
-    this.serialized = serialized;
-  }
-
-  renderServerSide(context?: Dict<Opaque>): void {
-    if (context) {
-      this.context = context;
-    }
-    this.setupServer();
-    this.populateHelpers();
-    this.element = this.env.getAppendOperations().createElement("div") as HTMLDivElement;
-    let template = expect(this.template, "Must set up a template before calling renderServerSide");
-    // Emulate server-side render
-    renderTemplate(this.env, template, {
-      self: new UpdatableReference(this.context),
-      parentNode: this.element,
-      dynamicScope: new TestDynamicScope(),
-      mode: "serialize"
-    });
-
-    this.takeSnapshot();
-    this.serialized = this.serialize();
-  }
-
-  serialize() {
-    let serializer = new SimpleDOM.HTMLSerializer(SimpleDOM.voidMap);
-    let serialized = serializer.serializeChildren(this.element);
-    return serialized;
-  }
-
-  renderClientSide(context?: Dict<Opaque>) {
-    if (context) {
-      this.context = context;
-    }
-    this.setupClient();
-    this.populateHelpers();
-    let { env } = this;
-    this.template = this.compile(this.rawTemplate);
-    this.element = env.getAppendOperations().createElement("div") as HTMLDivElement;
-    let template = expect(this.template, "Must set up a template before calling renderClientSide");
-    // Client-side rehydration
-    this.renderResult = renderTemplate(env, template, {
-      self: new UpdatableReference(this.context),
-      parentNode: this.element,
-      dynamicScope: new TestDynamicScope(),
-      mode: "rehydrate"
-    });
-  }
-
-  renderTemplate(template: Template<Opaque>): RenderResult {
-    this.template = template;
-    this.renderServerSide();
-    this.renderClientSide();
-    return this.renderResult!;
-  }
-}
-
 function normalize(oldSnapshot: NodesSnapshot, newSnapshot: NodesSnapshot, except: Array<Node>) {
   let oldIterator = new SnapshotIterator(oldSnapshot);
   let newIterator = new SnapshotIterator(newSnapshot);
@@ -826,7 +737,7 @@ export function renderTemplate(env: TestEnvironment, template: Template<Opaque>,
   return result;
 }
 
-function content(list: Content[]): string {
+export function content(list: Content[]): string {
   let out: string[] = [];
   let depth = 0;
